Implement a canvas-like pattern for creating attributes.
This makes it easy to get either a dictionary or OrderedCollection fo attributes and allows us to store the configuration in the attribute (so you can find out where it was defined).

It also makes the definition of attributes a little cleaner.

You now need to implement #addAttributes instead of #attributes.

Also updated configuration-related class comments.