The primary change here is to make SystemConfiguration>>localValueAt:ifAbsent: check only actual locally-overridden values. The rationale is that checking attributes (to get the default value) is slow and checking the local value is fast.
So we implement #inheritedValueAndSourceAt:do: to check all the local values first and if it cannot find one, to then search for the attribute and get the default value.

This actually gives little speedup in practice because the more common case tends to be the default value fore many configuration attributes but it doesn't seem to hurt across the average and will help a few edge cases.