Move WAAbstractPresenter>>renderContentOn: down to WADecoration. All Components override the method anyway so there's no need for them to inherit this version, which would end up rendering all children by default.
It seems much clearer to leave WAComponent abstract so it is clear that you need to implement #renderContentOn: for your class. If anyone doesn't like the abstractness, an empty #renderContentOn: could be added to WAPresenter. This would at least be obviously not a method you need to call with super.