I noticed that #decorationChainDo: only has one implementor and it only ever evaluates the block once, for either the outer decoration or the component itself. Furthermore, every sender just takes the value given to the block and calls a method on it.
Seems much simpler to just have a method return the appropriate object and then call a method on it.

I've chosen #outermost as a selector for the moment, though I don't exactly love it. I wanted to find something generic enough that it made sense having an implementation that returns self as high up the hierarchy as possible (currently its on WAPresenter but could even go up to WAPainter, possibly).

Anyway, the name can be changed if desired but the mechanism seems cleaner.