Since we've already stolen the idea of #renderContentOn: to come up with #describeOn:, we might as well steal the rendering context idea when determining the value of an attribute.
So we add WAAttributeContext and watch the code get less and less ugly.

Also do some renaming:

WAConfigurationValue -> WAAttributeExpression
WAOverrideValue -> WAValueExpression
WACollectionValue -> WAAddRemoveExpression

WAConfigurationAttribute -> WAAttributeDescription
because I don't think this object *is* the attribute, it merely describes it (particularly when you consider they are regenerated whenever the description is rebuilt).