Add an #expiredKeyHandlerClass configuration attribute for use by WARegistry.
Create WAExpiredKeyHandler for WARegistry, and WAExpiredSessionKeyHandler (also deletes cookies) for use with WAApplication. The session layer configuration was also updated to override the default value of the attribute to WAExpiredSessionKeyHandler.

Also updated the class comment for WARegistry to remove specific references to sessions (it is a more general implementation of a dynamically-keyed request dispatcher).

Also, set the parent of the not found handler in WAEntryPoint>>notFoundHandler