Gerhard Obermann reported a bug when clearing an AddRemoveExpressions from a configuration. It seems the default value was not restored properly.
The fix in Seaside-Development-Core-a.41 was not correct and has been reverted. This is the correct fix (I think), which is to ensure that WACollectionAttribute>>default returns a copy of its collection.  WAAddRemoveExpression>>determineValueWithContext: was also modified to ensure that it always makes a new set before doing the removals and additions.

This commit also includes a rename of the 'values' instance variable on WAConfigurationDescription to 'expressions', which should have been done when it started storing expressions.