Remove class-side helper methods on WACurrentRequestContext and WARequestContext. They were basically not being used and needed to be constantly added to as more instance methods were added.
Since most of the accesses go through one of the helpers on WAObject anyway, there's little benefit and it just adds confusion.

Also change a few pieces of code that were using them.

And add a WARequestContextNotFound error subclass which is signaled when calling WAObject>>requestContext with no current request context.

WAObject's #application and #session should maybe catch these and return nil to be compatible with Seaside 2.8 but http://code.google.com/p/seaside/issues/detail?id=140 was filed to discuss this.