Refactor to combine WANotFoundHandler, WAExpiredKeyHandler, WARedirectHandler, and WAExpireSessionKeyHandler into one object, WAResponseFactory.
WANotFoundHandler still remains as an actual handler because WADispatcher is written in such a way that it expects to always find a handler for a request, even if that is a "not-found" handler which just delegates the response generation to the factory.

This will pave the way for moving the rendering code off of WAResponse and onto a subclass of the factory.