Add WAExceptionHandler as a superclass of WAErrorHandler which can handle any kind of Exception.
Change the protocol so that the exception handler itself is given the responsibility of registering exceptions it wants to catch and dispatching them to appropriate methods.

WAErrorHandler, however, maintains the same protocol as before with the old methods marked as deprecated.

This follows discussion in http://code.google.com/p/seaside/issues/detail?id=160