Have WAErrorHandler catch halts (using platform support to get the halt class).
The default behaviour was a bit hard to select but I opted for displaying an internal error. This prevents hiding #halt calls in deployed code, and makes it obvious to the user that something is wrong if they are trying to develop but don't have the debugging error handler set. This default could arguably be changed.

This resolves http://code.google.com/p/seaside/issues/detail?id=160