Move the API deprecation stuff back onto Object and into the Platform package. We needed to be able to deprecate class-side methods and also class extension methods so it really needs to be on Object and we may need to deprecate interfaces on the platform class.
Rename #deprecatedApi: to #seasideDeprecatedApi:
Add #seasideDeprecatedApi:details: and an instvar to WADeprecatedApi to hold the detailed suggestions of how to change your code.
Remove the zero-parameter version of the method which relied on thisContext, which is apparently not portable. We just need to be explicit in all the senders now.