When I committed my change to move callback processing onto WACallbackRegistry I was, of course, confused because WACallbackRegistry wasn't actually holding onto the callbacks (it was storing them in the render context).
This would have been done initially so that the registry could be created holding onto the owner, however we no longer require the owner for the callbacks (this instance variable on WACallbackRegistry should have been removed as part of http://code.google.com/p/seaside/issues/detail?id=158 but was not) which means we can now have a single instance of the registry stored per render context and it can hold the callbacks itself.

This commit removes the "owner" instance variable on WACallbackRegistry and adds a "callbacks" instvar which holds a dictionary. WARenderingContext is modified to store a registry instead of a dictionary in its "callbacks" instvar. Users of both classes have been updated.