Move "parent" instvar of WARequestHandler to WAEntryPoint.
The instvar used to server double purpose as both the static parent of an entry point (set at registration time) and as a sort of request handler context stack (set during request handling on some error handlers). With the handler stack in WARequestContext that latter use is no longer required and the duality is confusing.

This moves the instvar down to WAEntryPoint, where it is always a static parent set once on a persistent handler. This led to deprecating #baseUrl on WASession, which I think makes sense anyway because it is the base url of the application that it returns in the first place.