Implement #cloneHandler on the RequestHandler hierarchy and use it in WACopyDispatcherPlugin.
For most request handlers we just want to #copy, but for Applications we want to make sure we then clear the cache. Otherwise we'd have a copy that had all the same sessions in it. Eww...

This was based on my assumption that in the general case of WAEntryPoint, we don't necessarily want to clear the cache in #postCopy. Somebody might have an EntryPoint subclass that they want to copy in some context and still have the registered request handlers in the cache. If this assumption is wrong, WAEntryPoint>>postCopy could be adjusted and this version reverted.