Merge and revert Seaside-Development-obi.3
While the fix in that version does solve the problem of the New Session link not working, it does so at the expense of not expiring the session at all (#unregistered is a notification, not an action, and does nothing on WASession).

That link needs to be fixed properly along with the rest of http://code.google.com/p/seaside/issues/detail?id=176 otherwise the omissions won't be noticed by whoever fixes that issue (since there will no longer be a sender of #expire there).

Alternatively, one could argue that "New Session" does not need to expire the session at all, in which case the call to #expire should be simply removed, rather than being replaced with #unregistered.