The existing cookie detection wasn't working when users had cookies disabled because it wasn't including the session key in the URL for the cookie check.
Instead we initialize the cookiesEnabled instvar to false, use the #useSessionCookie preference to determine whether to run the test, and if the test is successful we set the instvar to true. WASession>>url uses the instance variable to determine whether to include the session key in the URL so it will always be included until we are certain that we want to use cookies and that the user supports them.