WAResponse>>deleteCookie: was depending on having a current request context and an EntryPoint in that context.
It was being called from WAApplication when the first condition was not true and this was resulting in an error.

I moved the code from #deleteCookie: into #deleteCookie:path:, which WAApplication can use, passing in the cookie path, and had #deleteCookie: attempt to get the path from the request context and pass it into the version that takes the path.