http://code.google.com/p/seaside/issues/detail?id=229
When WAActionPhaseContinuation redirected to the render continuation, this was not wrapped with a handler for WAUnregisteredHandlerError. If a callback expired the session and then immediately triggered a render notification we would end up seeing that error when trying to get the URL for the new Continuation.

So we now add the handler in WASessionContinuation>>value and have it wrap the entired #handleRequest call.

Also renamed the methods that add the exception handlers to end with "Do:"