Get rid of WAView. Not that we have visitors implementing the presenter tree traversal, WAPresenter is no longer abstract.
I pulled the #children code up to WAPresenter and implemented it on WADecoration to return #next.

The implementation of #children on WADecoration actually ends up being slightly problematic because I actually have to ignore #next while traversing #children so that we don't call #acceptDecorated: on the owning component and start an infinite recursion. I may change this so that #next is not included in #children and we just reserve that hook for user-created custom decorations.