The goal of WARequestContext is to be a decoupling interface between request handlers and whatever wants the request handled. Rather than storing the server adaptor itself in the request context, it is therefore cleaner to make sure that the request context holds all information that needs to be exchanged. In fact, there is essentially no protocol common to all server adaptors that could be useful to a request handler and there were no users of WARequestContext>>server other than to get the codec.
This series of commits therefore changes WARequestContext to hold the codec that was used to decode the request and stops storing the server adaptor on the context.

If it turns out that people *do* want server-adaptor-specific information in their code, server adaptors could use custom subclasses of WARequestContext to provide it.