http://code.google.com/p/seaside/issues/detail?id=183
Deprecate WAApplication>>description and stop it getting its root component description. The method still exists on the class-side.

I also did not remove the class-side #description method from Components. It is still used by the example browser, though not anywhere much else. But the configuration application could decide to show it if desired.