As discussed on the mailing list, stop saving state at the end of the render phase. Snapshotted state should only be modified during callback processing and this has always been the case.
This halves the number of times we need to call #updateStates: on all the components.

I still haven't confirmed whether ajax callbacks should be doing "self states snapshot" or something like "states := self states copy snapshot; yourself" because JS code is still not working in this branch. The first will result in ajax callbacks modifying the (shared) state of the previous response continuation. This means that refreshing the page will reflect the new state which could be seen as a good thing but might result in weird behaviour since the next action callback would *also* share the same snapshot.