Move load script functionality from yucky implementation on RequestContext/RenderPhaseContinuation to HtmlDocument.
This removed the need for yucky #rememberLoadScripts: method as well.

I also fixed the two users of PrettyPrintedDocument (WAHalo and JQAllTests) so they work correctly.

One RSS test is now failing because it was using HtmlDocument for XML and there is now Html-specific behaviour (scripts) on there. I think this finally requires introduction of a WADocument class. I'll look at this next.

There's also one other issue remaining: the <script> block needs to *always* be output even if there are no load scripts and the RenderLoopContinuation is still attaching the "onload" attribute to the body.