http://code.google.com/p/seaside/issues/detail?id=179
Make sure all operations on the cache plugins happen within the cache's mutex. This is the simplest solution and can be further optimized if performance ever ends up being a problem. I found a couple of unprotected edge cases as well.

I also changed the protocol a bit to be more compatible with the Dictionary protocol and added method comments.