Fix two bugs in development error handlers.
One, WADebugErrorHandler was incorrectly generating a new WAResponse when you tried to resume a non-resumable exception.

Two, WAWalkbackErrorHandler is doing a call. When it comes back from the call, the current request context has changed. WAExceptionHandler stores the request context, though, because they are useable without RenderLoop. Since this particular class depends on RenderLoop and does a #call:, we need to update the value of the instance variable with the new current request once the call returns.