http://code.google.com/p/seaside/issues/detail?id=31
There are no senders of WAStaticHtmlCanvas. The only thing it did was allow you to hardcode a Document different from that in the RenderContext. Presumably this made sense at some point in the past but you can just make a new RenderContext (or copy one you already have) or change the document on the RenderContext. No need for a special renderer to do this.