Remove dependency of WAExceptionFilter on Session and move it into the Core package. Have it store its own configuration.
Still no UI for actually configuring it but I'll do that next. In the meantime, Applications added with WAAdmin will have a new named configuration added to the filter so that can be used to change the exception handler for *all* applications.

You'll need to clear  your config cache and reload your applications after loading this.