Split WAAdmin into its own package Seaside-Tools-Core.
Some of the methods on there will need to get moved back into Seaside-Environment again but I still need to sort out the details. The idea is that the Environment will use WAAdmin to install all the applications we want installed by default but WAAdmin itself is just a tool.