Address http://code.google.com/p/seaside/issues/detail?id=367 by removing WAStartableServerAdaptor.
After discussing with Lukas, the existing interface was about the best we could come up with so I implemented something similar as class extensions from Tools-Core. These methods will work as long as you do not have multiple adaptors of the class you are trying to work with registered, in which case it will error.

The behaviour from WAStartableServerAdaptor was pushed onto WAKom so that those compatibility classes still maintain the exact behaviour of WAKom in 2.8 (i.e. they store a "default" instance in a class instance variable).

I also fixed a bug introduced into WAKomEncoded with the codec refactoring which meant that it was registering an adaptor with a WANullCodec.