http://code.google.com/p/seaside/issues/detail?id=406
Move tests for the registration of default handlers into the tests for the packages that actually define the handlers.

This may not ultimately be the correct solution if we move to a declaritive mechanism of defining these things and have them loaded by a single package but it is correct for the current behaviour and gets these long-failing tests passing.

Note that Seaside-Tests-Examples is a new package.