http://code.google.com/p/seaside/issues/detail?id=405
Start storing the query fields in the Request URL and adjust accessors so that #queryFields returns those, #postFields returns the POST fields, and #fields returns a merge of the two.

This is the simplest, least invasive implementation of this I could come up with. I still think we need to put some serious thought into how field merging and so on works.