Properly fix http://code.google.com/p/seaside/issues/detail?id=544
I introduce a GROrderedMultiMap which specifically allows duplicate keys (GRSmallDictionary no longer does). This is used to implement WARequestFields, which is used throughtout for storing GET and POST fields. #at: will return the first matching field and #allAt: will return all matching values.

Various renamings, fixes to callbacks, and to unit and functional tests.

I also had to fix the Swazoo and Comanche server adaptors to correctly create WARequestFields instances for the POST fields and to correctly include the raw POST body in all cases (but it seems to be a ByteArray in Swazoo's case and a ByteString or String for Comanche... seems not ideal)