- fixed encoding in javascript strings with html tags (thanks goes to John Toohey)
I reduce the code to :-

 $("#markettable").prepend(
   "<script></script>"
  );

but this still gives an error.

However, adding escaping the final forward / in the script tag fixes this. I found a lot of reports of this issue on the net, and the consensus seems to be to always escape a forward / whilst emitting JS to the browser.