http://code.google.com/p/seaside/issues/detail?id=610
Using the welcome screen component creator on a fresh image causes the image to to request an author name before the component is created. This is clumsy because the user has to enter a username in the Smalltalk environment, but there is no warning of this in the browser.

Fixed so the welcome component creator checks if an author name has been created in the image. If not the welcome application prompts for author name.